% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis.STRUCTURE.popgen.R
\name{fis.STRUCTURE.popgen}
\alias{fis.STRUCTURE.popgen}
\title{Compare Fis values between population clusters}
\usage{
fis.STRUCTURE.popgen(qmatrix, allele.dat, nclusters, nreps, NA.symbol,
  samp.snps, nsamp, ncores)
}
\arguments{
\item{qmatrix}{Dataframe. The STRUCTURE qmatrix, with individuals in column 'ind'}

\item{allele.dat}{Dataframe/matrix containing '/' separated allele data,
with rownames representing individuals}

\item{nclusters}{(optional) Integer. The number of clusters identified by STRUCTURE}

\item{nreps}{Integer. The number of times to repeat estimations. Default is 100}

\item{NA.symbol}{The code used to represent missing allele data}

\item{samp.snps}{Logical. If TRUE, a specified number of SNPs will be randomly sampled
in each iteration to account for uncertainty and speed computations. Default is FALSE}

\item{nsamp}{(optional) Integer. The number of SNPs to sample during each iteration}

\item{ncores}{Integer. The number of cores to split job across. Default is 1 (no parallelisation)}
}
\value{
Matrix of pairwise comparison confidence intervals
}
\description{
This function assigns individuals to clusters using STRUCTURE
assignment probabilities provided in the qmatrix. It then calculates
individual Fis values and compares cluster-level Fis using an ANOVA.
}
