% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbreed.STRUCTURE.popgen.R
\name{inbreed.STRUCTURE.popgen}
\alias{inbreed.STRUCTURE.popgen}
\title{Subsample SNPs and estimate individual inbreeding coefficients}
\usage{
inbreed.STRUCTURE.popgen(allele.dat, qmatrix.prob, samp.snps, nreps, ncores,
  nsamp, NA.symbol, qmatrix = NULL, ind.populations = NULL,
  nclusters = NULL)
}
\arguments{
\item{allele.dat}{Dataframe/matrix containing '/' separated allele data,
with rownames representing individuals}

\item{qmatrix.prob}{(optional) Logical. If TRUE, a supplied STRUCTURE qmatrix
will be used to randomly assign individuals to clusters based on assignment probabilities.
Default is FALSE}

\item{samp.snps}{Logical. If TRUE, a specified number of SNPs will be randomly sampled
in each iteration to account for uncertainty and speed computations. Default is FALSE}

\item{nreps}{Integer. The number of times to repeat the subsampling and
estimate inbreeding. Default is 100}

\item{ncores}{(optional) Integer. The number of cores to split job across.
Default is 1 (no parallelisation)}

\item{nsamp}{(optional) Integer. The number of SNPs to sample during each iteration}

\item{NA.symbol}{The code used to represent missing allele data}

\item{qmatrix}{(optional) The STRUCTURE qmatrix, with individuals in column 'ind'. To be used
if qmatrix.prob = TRUE}

\item{ind.populations}{(optional) Matrix. Population assignments for individuals,
with rownames representing individuals}

\item{nclusters}{(optional) Integer. The number of clusters identified by STRUCTURE}
}
\value{
A list of two elements:\cr\cr
lin.coefs:   Tukey's population comparisons from repeated ANOVAs comparing individual
inbreeding estimates across populations.\cr\cr
raw.vals:   summaries of the raw inbreeding estimates for each individual
}
\description{
This function assigns individuals to known (i.e. provided) populations
or to clusters using STRUCTURE assignment probabilities provided in the qmatrix.
It then calculates inbreeding coefficients for individuals and compares
cluster-level inbreeding with an ANOVA.
}
