% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.metrics.STRUCTURE.popgen.R
\name{obs.metrics.STRUCTURE.popgen}
\alias{obs.metrics.STRUCTURE.popgen}
\title{Compute observed heterozygosities and richness metrics for STRUCTURE clusters}
\usage{
obs.metrics.STRUCTURE.popgen(qmatrix, allele.dat, nclusters, nreps, NA.symbol,
  samp.snps, nsamp, ncores)
}
\arguments{
\item{qmatrix}{Dataframe. The STRUCTURE qmatrix, with individuals in column 'ind'}

\item{allele.dat}{Dataframe. Raw allele data (sep = '/')}

\item{nclusters}{Integer. The number of clusters identified by STRUCTURE}

\item{nreps}{Integer. The number of times to repeat estimations. Default is 100}

\item{NA.symbol}{The code used to represent missing allele data}

\item{samp.snps}{Logical. If TRUE, a specified number of SNPs will be randomly sampled
in each iteration to account for uncertainty and speed computations. Default is FALSE}

\item{nsamp}{(optional) Integer. Number of SNPs to sample during each iteration. If
samp.snps = TRUE and nsamp is not provided, default is to sample 50% of the loci
(i.e. ncol(allele.dat)/2) in each iteration}

\item{ncores}{(optional) Integer. Number of cores to split job across for parallel
computing. Default is 1 (no parallelisation)}
}
\value{
Matrix of mean and 95% quantiles of estimated allele frequencies per cluster
}
\description{
This function randomly assigns individuals to clusters using STRUCTURE
assignment probabilities provided in the qmatrix. It then uses resampling
to generate distributions of estimated heterozygosities and allelic richness metrics
per locus, per cluster.
}
