% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRUCTURE.popgen.R
\name{STRUCTURE.popgen}
\alias{STRUCTURE.popgen}
\title{Calculate population genetics parameters}
\usage{
STRUCTURE.popgen(qmatrix = NULL, allele.dat, nclusters, stat, NA.symbol,
  samp.snps, nsamp)
}
\arguments{
\item{qmatrix}{Dataframe. The STRUCTURE qmatrix, with individuals in column 'ind'}

\item{allele.dat}{Dataframe. Raw allele data (sep = '/')}

\item{nclusters}{Integer. The number of clusters identified by STRUCTURE}

\item{stat}{The population genetics statistic to return. Must be one of the following: 'ind.dist',
'summary', 'allele.freqs',
'heterozygosity', 'richness','fis','fst'. \cr\cr
Note that statistics 'ind.dist','summary' and 'heterozygosity' do not require a qmatrix, while
statistics 'allele.freqs', 'richness','fis' and 'fst' must have a supplied qmatrix.}

\item{NA.symbol}{The code used to represent missing allele data}

\item{samp.snps}{Logical. If TRUE, a specified number SNPs will be randomly sampled
in each iteration to account for uncertainty and speed computations. Default is FALSE}

\item{nsamp}{(optional) Integer of SNPs to sample during each iteration}
}
\value{
Summaries based on the chosen population genetic statistic
}
\description{
This function randomly assigns individuals to clusters using STRUCTURE
assignment probabilities provided in the qmatrix. The qmatrix must have
column 'ind', which stores individual ids.
}
